Fix DataFrameSuiteBase for Spark 2.0
