+++ src/main/2.0/scala/com/holdenkarau/spark/testing/JavaDataFrameSuiteBase.scala
@@ -1,0 +1,4 @@
+package com.holdenkarau.spark.testing
+
+class JavaDataFrameSuiteBase extends SharedJavaSparkContext with DataFrameSuiteBaseLike with JavaTestSuite {
+}
