+++ src/main/pre-2.0/scala/com/holdenkarau/spark/testing/JavaDataFrameSuiteBase.scala
